+++ src/game/collision/Collision.java
@@ -140,7 +140,7 @@ public class Collision {
     private static Vector3d fPos;
     private static double x;
     private static double z;
-    private static int cachedBlock;
+    private static byte cachedBlock;
     private static final Vector3d cachedPos = new Vector3d(0d,0d,0d);
 
     //sneaking stuff
@@ -217,8 +217,8 @@ public class Collision {
         return binaryReturn;
     }
 
-    private static boolean sneakCollideYNegative(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, float width, float height, boolean onGround, int blockID){
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+    private static boolean sneakCollideYNegative(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, float width, float height, boolean onGround, byte blockID){
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX,blockPosY,blockPosz,thisBlockBox);
 
@@ -385,8 +385,8 @@ public class Collision {
 
     //a simple way to check if an object is in the water, only done on x and z passes so you can't stand
     //next to water and get slowed down
-    private static void detectIfInWater(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, float width, float height, int blockID){
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+    private static void detectIfInWater(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, float width, float height, byte blockID){
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX,blockPosY,blockPosz,thisBlockBox);
 
@@ -400,8 +400,8 @@ public class Collision {
         }
     }
 
-    private static boolean collideYNegative(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, Vector3f inertia, float width, float height, boolean onGround, int blockID){
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+    private static boolean collideYNegative(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, Vector3f inertia, float width, float height, boolean onGround, byte blockID){
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX,blockPosY,blockPosz,thisBlockBox);
             //this coordinate is not within enough distance to get affected by floating point precision
@@ -417,8 +417,8 @@ public class Collision {
         return onGround;
     }
 
-    private static void collideYPositive(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, Vector3f inertia, float width, float height, int blockID){
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+    private static void collideYPositive(int blockPosX, int blockPosY, int blockPosz, byte rot, Vector3d pos, Vector3f inertia, float width, float height, byte blockID){
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX, blockPosY, blockPosz,thisBlockBox);
             //this coordinate is not within enough distance to get affected by floating point precision
@@ -432,9 +432,9 @@ public class Collision {
 
     //TODO ----------------------------------XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
 
-    private static void collideX(int blockPosX, int blockPosY, int blockPosZ, byte rot, Vector3d pos, Vector3f inertia, Vector3d oldPos, float width, float height, int blockID){
+    private static void collideX(int blockPosX, int blockPosY, int blockPosZ, byte rot, Vector3d pos, Vector3f inertia, Vector3d oldPos, float width, float height, byte blockID){
         //run through X collisions
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
 
@@ -473,7 +473,7 @@ public class Collision {
 
             //collide X negative
             if (blockPosX + 0.5d <= pos.x) {
-                for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+                for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
                     setAABB(pos.x, pos.y, pos.z, width, height);
                     setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
                     if (isWithin() && BlockBoxGetRight() > AABBGetLeft() && BlockBoxGetRight() - AABBGetLeft() < 0.01d) {
@@ -485,7 +485,7 @@ public class Collision {
 
             //collide X positive
             if (blockPosX + 0.5d >= pos.x) {
-                for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+                for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
                     setAABB(pos.x, pos.y, pos.z, width, height);
                     setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
                     if (isWithin() && BlockBoxGetLeft() < AABBGetRight() && BlockBoxGetLeft() - AABBGetRight() > -0.01d) {
@@ -500,9 +500,9 @@ public class Collision {
 
     //TODO ----------------------------------ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ
 
-    private static void collideZ(int blockPosX, int blockPosY, int blockPosZ, byte rot, Vector3d pos, Vector3f inertia, Vector3d oldPos, float width, float height, int blockID){
+    private static void collideZ(int blockPosX, int blockPosY, int blockPosZ, byte rot, Vector3d pos, Vector3f inertia, Vector3d oldPos, float width, float height, byte blockID){
         //run through Z collisions
-        for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+        for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
             setAABB(pos.x, pos.y, pos.z, width, height);
             setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
 
@@ -541,7 +541,7 @@ public class Collision {
 
             //collide Z negative
             if (blockPosZ + 0.5d <= pos.z) {
-                for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+                for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
                     setAABB(pos.x, pos.y, pos.z, width, height);
                     setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
                     if (isWithin() && BlockBoxGetBack() > AABBGetFront() && BlockBoxGetBack() - AABBGetFront() < 0.01d) {
@@ -553,7 +553,7 @@ public class Collision {
 
             //collide Z positive
             if (blockPosZ + 0.5d >= pos.z) {
-                for (double[] thisBlockBox : getBlockShape(blockID, rot)) {
+                for (float[] thisBlockBox : getBlockShape(blockID, rot)) {
                     setAABB(pos.x, pos.y, pos.z, width, height);
                     setBlockBox(blockPosX, blockPosY, blockPosZ, thisBlockBox);
                     if (isWithin() && BlockBoxGetFront() < AABBGetBack() && BlockBoxGetFront() - AABBGetBack() > -0.01d) {
@@ -575,7 +575,7 @@ public class Collision {
                  AABBGetBack() < BlockBoxGetFront());
     }
 
-    private static int detectBlock(Vector3d flooredPos){
+    private static byte detectBlock(Vector3d flooredPos){
         return getBlock((int)flooredPos.x, (int)flooredPos.y, (int)flooredPos.z);
     }
